+++ src/Exceptions/RepositoryException.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace PPSpaces\Exceptions;
+
+use Exception;
+
+class RepositoryException extends Exception
+{
+    /**
+     * Report the exception.
+     *
+     * @return void
+     */
+    public function report()
+    {
+        //
+    }
+
+    /**
+     * Render the exception into an HTTP response.
+     *
+     * @param  \Illuminate\Http\Request
+     * @return \Illuminate\Http\Response
+     */
+    public function render($request)
+    {
+        // return response(...);
+    }
+}